backend-web-produksi/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
prisma==0.12.0
python-dotenv==1.0.1
python-multipart==0.0.9
